Replace dbms connection attribute
Current code is written for ibm_db backend, replaced call to
get the dbms name and dbms version to the pyodbc method.
Additionally, in the DB2DDLCompiler, the method to check if the
unique constraint is nullable checks the version of Db2 LUW and
that is changed to a IBM i Db2 version.

Fixes: #37