Refactor code to pass sqlalchemy test cases
Separated backend and base again and swapped PyODBCConnector to
be inherited first as the functions needed to be inherited from
that class for the package to function properly. Additionally,
the unimplimented abstract classes were causing errors in the
tests so they will be removed in this branch until they are
implemented. Lastly reverted a few past changes as they were not
supposed to be made and caused errors in the testing.