Testing setup (#51)
Updated various pieces to run the standard SQLAlchemy test suite,
adjusting some testing requirements to disable/enable tests for IBM i
as appropriate.

Since there is currently no way to run the tests in GitHub Actions,
that section of the action is removed. Previously this was a no-op
because there was no test suite defined to run.