Replace dbms connection attribute (#39)
The version has been cached using the pyodbc method in #76 already, so
the old version caching code can be removed. Likewise, the old database
name caching code can be removed since this adapter only supports the
IBM i database.

Additionally, in the DB2DDLCompiler, the method to check if the
unique constraint is nullable checks the version of Db2 LUW. Since 
this feature is not supported on IBM i, these code paths were also 
removed. 

Fixes: #37