Disable Extended Dynamic support
The driver defaults to enabling Extended Dynamic, as it allows caching
and re-use of queries on the server side. Ideally, this is a boon, but
the way the ODBC driver sets things up is problematic for use within
interpreted languages or middleware. The driver supports both
DefaultPkgLibrary and DefaultPackage options to configure how packages
are used with DefaultPkgLibrary specifying the library to hold the
package and DefaultPackage configures how the package should be used,
but no way to control the name of the package. Instead, the package name
will be set based on the name of the calling application, which for any
Python application will be "python", "python3", or "python3.XX". This
means all SQLAlchemy applications would share the same package, even if
they use different library lists or default schemas, which can result in
unexpected behavior from queries using unqualified object names using
the wrong object due to the cached access plan in the package.

To work around this problem, we disable Extended Dynamic support for
now. If necessary, a connection option could be added that allows
setting the package name and then we could allow users to set both the
package name and library.