Make option mapping code more generic
This will make it easier to add more options in the future. In addition,
we can easily work around the 07.01.025 driver bug which only supports
case-sensitive connection options and is currently broken when using the
pyodbc/SQLAlchemy default options.

Also, convert boolean values to '1'/'0', since otherwise they will get
converted to 'True'/'False', which the driver does not understand.