Set dialect name to ibmi
The dialect name is not supposed to be the package name, but should be
the database name, eg. from upstream: mysql, postgresql, oracle, etc
While we could use "db2" here, ibm_db_sa doesn't use this (we actually
inherited this misconstrued name when we forked from them and set it to
our new package name) and we are different enough from Db2 LUW that I
think it makes sense. We can always re-evaluate prior to 1.0.

With this, also skip tests (instead of passing) using the testing.skip
decorator and passing our database name.