Remove "data" from reserved word list
This causes errors running HasTableTest, since the data column gets
quoted and then the view can't be created because it's unquoted:

    CREATE VIEW vv AS SELECT id, data FROM test_table

This test isn't a problem on any other database because AFAICT on all
other providers SQLAlchemy ships (other than Oracle), quoted names are
not case-sensitive and since data is not a reserved word on Oracle, the
test is just fine on Oracle too.

The upstream test could get changed to use a more unlikely to be
reserved name or use the APIs to create the views, but I haven't
analyzed the test enough to know why they're using a direct query like
that. Since although data is listed as a reserved word, it doesn't seem
to cause any problems not quoting it, for now just comment it out.