Pin lint action to Python 3.9 for now
Using 3.x broke now that it's using 3.12 due to pyodbc incompatibilities
as certain deprecated APIs used by pyodbc were removed. With the release
of pyodbc 5.0.0, Python 3.12 is now supported but we would need to add
custom version handling based on Python version.

For now, pin this to Python 3.9, the most recent version we build PRs
with. This should also get updated to support newer Python versions...