Implement LIMIT / OFFSET properly
LIMIT is kinda strange on IBM i. It's not implemented as a separate
clause like OFFSET is, but as an alternate form of FETCH clause. It's
there for some compatibility with other databases, but it's not 100%
compatible. The compatibility syntax for LIMIT ... OFFSET is also
*different* than the OFFSET clause, which is confusing. Due to this,
each combination would need it's own code path. Alternatively, we can
use Db2 OFFSET syntax with the FETCH syntax (SQL reference describes how
LIMIT syntax maps to FETCH) in a consistent way, so that's what we do.
With this, we no longer need our custom visit_select function to handle
limit and offset, since that should be handled already.

In addition, we no longer need to override the broken LimitOffsetTest
cases. We can also open some requirements for features we support to
enable more of the FetchLimitOffsetTest test cases to run instead of
skipped.