Cast NULL to type and always call parent visit_cast
We can't use a NULL constant/literal in a parameter list without a type
or we'll get SQL0206 - Column or global variable NULL not found.
We can work around this by casting to a type, but at this point we don't
know what the type was, and when using the null() function, there will
not be a type anyway, so we pick an arbitrary type of INTEGER which is
most compatible with other types other than BLOB, XML, and some others.

In addition, always call super().visit_cast, not just for some types. I
don't know why this was ever this way, but it certainly breaks custom
types and likely other built-in types as well.

With this, CastTypeDecoratorTest tests, ExistTest tests, and the
remaining NumericTest tests now pass.