Remove overridden get_select_precolumns
This function doesn't need to do anything special that the base function
doesn't already do.

The only extra thing it's doing is some special handling if
self._distinct is a string. This seems to be some ancient, deprecated
behavior in the MySQL dialect which was probably copied in unkowningly.
Because we override this function, we don't emit the warning that we are
silently ignoring DISTINCT ON, which the DistinctOnTest expects. Since
we don't have any existing applications, we don't need any compatibility
deprecation like MySQL.

See SQLAlchemy commit 07d6d211f23f1d9d1d69fd54e8054bccd515bc8c and
https://github.com/sqlalchemy/sqlalchemy/issues/4002 for more details
about the upstream change.