Ensure names starting with 0-9 are quoted
The illegal_initial_characters set is checked to see if the first
character is in this to require quoting, but the first character will
never match an integer since they're different types.

We inherited this bug from ibm_db_sa. It seems SQLAlchemy had this same
bug introduced in 2009 in commit a43a0e8b68 and was fixed upstream in
2010 in c5e29f0eed. Perhaps ibm_db_sa inherited the bug by basing off
that codebase?

This fixes two DifficultParametersTest test cases.