Remove overridden visit_join
In SQLAlchemy 1.4, linting was added to warn about cartesian products:
https://docs.sqlalchemy.org/en/14/changelog/migration_14.html#change-4737
However, our visit_join doesn't know anything about the linter so we get
linter warnings when we shouldn't. We could update this function to do
so, but AFAICT we don't need to override this anyway.

The base compiler will generate the following joins:
- FULL OUTER JOIN (join.full)
- LEFT OUTER JOIN (join.isouter)
- JOIN

Db2 for i supports the following joins:

- (INNER) JOIN
- LEFT (OUTER) JOIN
- RIGHT (OUTER) JOIN
- FULL (OUTER) JOIN
- (LEFT) EXCEPTION JOIN
- RIGHT EXCEPTION JOIN
- CROSS JOIN

What the base compiler generates seems perfectly fine for Db2, so
there's no need to even override this function. Maybe in ancient times
this was needed? (the most recent refactor of this function in ibm_db_sa
was for SQLAlchemy 0.8 support)

This fixes 3 JoinTest test cases.