Fix ComponentReflectionTest failures on SA 1.4
Upstream class replaced the test_get_unique_constraints and
test_get_unique_constraints_with_schema test wrappers around
_test_get_unique_constraints with a single test method using
@testing.combinations. This caused our override to fail on SQLAlchemy
1.4. Due to changes in test harness, it will be hard to maintain this
test for both without duplicating it so instead, we just skip it.

In addition, the test_get_table_names combinations where include_views
was true were failing due to an upstream bug. Because we did not
override the views requirement to say we supported views, the views were
not created, but these view-related tests were not skipped due to
missing exclusion. Since we support views, we should be opening that
requirement anyway.