Overhaul how character types are compiled
Modeled after how the SQLAlchemy upstream does it, use a common function
to render the string types, handling type name, length, CCSID, and other
common attributes.

Remove LONGVARCHAR and LONGVARGRAPHIC types and supporting functions.
These are not real types in Db2 for i, but rather special syntax for
VARCHAR(n) and LONGVARGRAPHIC(n), where n is determined by the amount of
space left over in the row after all the other columns. This limits it
to 32716 bytes max, and for each long column in the row, the remaining
space is split evenly. This length cap is not a good fit for UnicodeText
which is supposed to be "unbounded". Additionally, this syntax is
considered deprecated and not recommended for use. UnicodeText will now
use a CLOB instead of LONGVARCHAR.

For LOB types, when length is not specified, default to their max length
instead of default length (1M). This matches what other databases do as
equivalent types on MySQL, PostgreSQL, MSSQL do not support a length,
but instead have an implicit max length between 1-4G (depending on the
database). And while Oracle does have a length for its CLOB type, it
defaults to its max of 2G. This also matches how SQLAlcheny describes
these types as having "unbounded" length.

Finally, move dialect-specific types together into a separate section.
DOUBLE has been added to SQLAlchemy 2.0, so we won't need this once we
drop support for 1.4 and earlier, but that's a ways off.