Add custom NCHAR, NVARCHAR, NCLOB handlers
The base versions of these would actually work, provided a collation
wasn't specified. This gives us the ability to handle that condition in
the future. It also allows us to handle NCLOB like we do for other LOBs,
by supporting a length argument and defaulting to its max length when
none is specified.