Clean up dialect options
preexecute_sequences was changed to preexecute_pk_sequences in
SQLAlchemy 0.4 with a backwards compatibility alias. That alias was
removed in SQLAlchemy 0.5 and preexecute_pk_sequences is no longer
referenced outside of the changelogs so remove it.

supports_unicode_binds and supports_unicode_statements default to True
in the default dialect when run under Python 3. Since we don't support
Python 2, just remove them (they were also removed in SQLAlchemy 2.0
anyway).

supports_sane_rowcount should be able to be enabled since the IBM i
Access driver will return row counts for UPDATE and DELETE statements.
Since the default dialect enables it by default, we can simply remove
the override.

supports_sane_rowcount_returning is enabled by default for the pyodbc
connector, so we also enable it. Since we don't support RETURNING, this
should be the same as supports_sane_rowcount. Since the pyodbc connector
enables it, we do to.

supports_char_length was at one point used to switch between VARCHAR(n)
and LONG VARCHAR on databases which didn't support VARCHAR with a
length. ibm_db_sa dropped this behavior in commit 1c4ccd9e1, but the
supports_char_length attribute was left. Since this is no longer used,
remove it.

Since we only support a single driver, just hard code the name and
remove pyodbc_driver_name. This was inherited from the pyodbc connector
code, but that is designed to be generic.

Neither two_phase_transactions nor savepoints seem to be dialect attributes,
but rather requirement properties. Maybe at one time they were, but they
no longer are so move them to requirements. Since two_phase_transactions
defaults off anyway, we don't need to do anything for it.