Enable SQLAlchemy 1.4+ statement caching
We need to ensure that any literals values put in a statement are not
embedded during the initial compilation stage. The example given in the
docs [1] is for LIMIT / OFFSET, but we process the limit and offset
values normally, not by inserting them and the FetchLimitOffsetTest
class will catch this as well.

We also have issues with selecting literal() values. For the most part,
we handle this by casting the parameter marker, but for Decimal objects,
we need to insert them as literals. For SQLAlchemy 1.4 and up, we use
the render_literal_execute post-processing facility, but for SQLAlchemy
we use literal_binds. Using literal_binds does embed it directly in the
statement, but on 1.3 and below there's no statement cache anyway.